+++ cpp/tests/fuzz/fuzzmaster.cpp
@@ -77,7 +77,7 @@ private:
 
 extern "C" int LLVMFuzzerTestOneInput(const uint8_t* Data, size_t Size)
 {
-    ser4cpp::rseq_t buffer(Data, static_cast<uint32_t>(Size));
+    ser4cpp::rseq_t buffer(Data, Size);
 
     opendnp3::MasterParams config;
     MasterTestObject t(config);
